+++ Classes/Hooks/PageRenderer/SchemaMarkupInjection.php
@@ -19,7 +19,6 @@ use TYPO3\CMS\Core\Cache\Frontend\FrontendInterface;
 use TYPO3\CMS\Core\Configuration\ExtensionConfiguration;
 use TYPO3\CMS\Core\EventDispatcher\EventDispatcher;
 use TYPO3\CMS\Core\Page\PageRenderer;
-use TYPO3\CMS\Core\Utility\ExtensionManagementUtility;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
 use TYPO3\CMS\Core\Utility\VersionNumberUtility;
 use TYPO3\CMS\Extbase\Object\ObjectManager;
@@ -103,14 +102,8 @@ final class SchemaMarkupInjection
 
     private function shouldEmbedMarkup(): bool
     {
-        $embedMarkup = true;
-
-        if (ExtensionManagementUtility::isLoaded('seo')) {
-            $embedMarkup = $this->controller->page['no_index'] === 0;
-        }
-
         $objectManager = GeneralUtility::makeInstance(ObjectManager::class);
-        $event = new ShouldEmbedMarkupEvent($this->controller->page, $embedMarkup);
+        $event = new ShouldEmbedMarkupEvent($this->controller->page, true);
 
         if (VersionNumberUtility::convertVersionNumberToInteger(TYPO3_branch) >= 10000000) {
             $eventDispatcher = $objectManager->get(EventDispatcher::class);